+++ web-app/admin/src/app/admin/admin-layers/imagery-layer-settings/imagery-layer-settings.component.scss
@@ -1,0 +1,361 @@
+.imagery-settings-container {
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+}
+
+.form-field {
+    display: flex;
+    flex-direction: column;
+    gap: 0.5rem;
+}
+
+.field-label {
+    font-weight: 600;
+    font-size: 0.9rem;
+    color: #333;
+}
+
+.form-input {
+    padding: 0.5rem;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+    font-size: 0.9rem;
+}
+
+.form-input:focus {
+    outline: none;
+    border-color: #4285f4;
+    box-shadow: 0 0 0 2px rgba(66, 133, 244, 0.1);
+}
+
+.radio-group {
+    display: flex;
+    gap: 1.5rem;
+    flex-wrap: wrap;
+}
+
+.radio-label {
+    display: flex;
+    align-items: center;
+    gap: 0.5rem;
+    cursor: pointer;
+    user-select: none;
+}
+
+.radio-label input[type="radio"] {
+    cursor: pointer;
+}
+
+.checkbox-label {
+    display: flex;
+    align-items: center;
+    gap: 0.5rem;
+    cursor: pointer;
+    user-select: none;
+}
+
+.checkbox-label input[type="checkbox"] {
+    cursor: pointer;
+}
+
+.action-button {
+    padding: 0.6rem 1.2rem;
+    border: 1px solid #ccc;
+    border-radius: 4px;
+    background-color: #fff;
+    color: #333;
+    font-size: 0.9rem;
+    cursor: pointer;
+    display: inline-flex;
+    align-items: center;
+    gap: 0.5rem;
+    transition: background-color 0.2s, border-color 0.2s;
+}
+
+.action-button:hover:not(:disabled) {
+    background-color: #f5f5f5;
+    border-color: #999;
+}
+
+.action-button:disabled {
+    opacity: 0.5;
+    cursor: not-allowed;
+}
+
+.btn-secondary {
+    background-color: #6c757d;
+    color: white;
+    border-color: #6c757d;
+}
+
+.btn-secondary:hover:not(:disabled) {
+    background-color: #5a6268;
+    border-color: #545b62;
+}
+
+.wms-section,
+.imagery-preview-section {
+    display: flex;
+    flex-direction: column;
+    gap: 1rem;
+}
+
+.wms-error {
+    padding: 0.75rem;
+    background-color: #fff3cd;
+    border: 1px solid #ffc107;
+    border-radius: 4px;
+    color: #856404;
+    display: flex;
+    align-items: center;
+    gap: 0.5rem;
+}
+
+.wms-server-info {
+    padding: 1rem;
+    background-color: #f8f9fa;
+    border: 1px solid #dee2e6;
+    border-radius: 4px;
+}
+
+.wms-server-info h4 {
+    margin: 0 0 0.5rem 0;
+    font-size: 1.1rem;
+    color: #333;
+}
+
+.wms-abstract {
+    margin-bottom: 0.5rem;
+    color: #666;
+    font-size: 0.9rem;
+}
+
+.wms-version {
+    margin-bottom: 0.5rem;
+    font-weight: 600;
+    font-size: 0.85rem;
+}
+
+.wms-contact {
+    font-size: 0.85rem;
+    color: #555;
+}
+
+.wms-contact div {
+    margin-bottom: 0.25rem;
+}
+
+.wms-contact a {
+    color: #4285f4;
+    text-decoration: none;
+}
+
+.wms-contact a:hover {
+    text-decoration: underline;
+}
+
+.wms-layers-section,
+.wms-other-layers-section {
+    display: flex;
+    flex-direction: column;
+    gap: 0.75rem;
+}
+
+.wms-layers-section h5,
+.wms-other-layers-section h5 {
+    margin: 0;
+    font-size: 0.95rem;
+    font-weight: 600;
+    color: #333;
+}
+
+.warning-text {
+    color: #856404;
+}
+
+.wms-layers-list {
+    max-height: 300px;
+    overflow-y: auto;
+    border: 1px solid #dee2e6;
+    border-radius: 4px;
+    padding: 0.5rem;
+    background-color: #fff;
+}
+
+.wms-layer-item {
+    display: flex;
+    align-items: flex-start;
+    gap: 0.75rem;
+    padding: 0.75rem;
+    cursor: pointer;
+    border-bottom: 1px solid #f0f0f0;
+    transition: background-color 0.2s;
+}
+
+.wms-layer-item:last-child {
+    border-bottom: none;
+}
+
+.wms-layer-item:hover:not(.disabled) {
+    background-color: #f8f9fa;
+}
+
+.wms-layer-item.disabled {
+    opacity: 0.6;
+    cursor: not-allowed;
+}
+
+.wms-layer-item input[type="checkbox"] {
+    margin-top: 0.2rem;
+    cursor: pointer;
+}
+
+.wms-layer-item.disabled input[type="checkbox"] {
+    cursor: not-allowed;
+}
+
+.layer-info {
+    display: flex;
+    flex-direction: column;
+    gap: 0.25rem;
+    flex: 1;
+}
+
+.layer-title {
+    font-weight: 600;
+    font-size: 0.9rem;
+    color: #333;
+}
+
+.layer-abstract {
+    font-size: 0.85rem;
+    color: #666;
+    line-height: 1.4;
+}
+
+.wms-advanced-section,
+.wms-capabilities-section {
+    margin-top: 0.5rem;
+}
+
+.wms-advanced-section h4,
+.wms-capabilities-section h4 {
+    margin: 0 0 0.75rem 0;
+    font-size: 0.95rem;
+}
+
+.wms-advanced-section h4 a,
+.wms-capabilities-section h4 a {
+    color: #4285f4;
+    text-decoration: none;
+    display: flex;
+    align-items: center;
+    gap: 0.5rem;
+}
+
+.wms-advanced-section h4 a:hover,
+.wms-capabilities-section h4 a:hover {
+    text-decoration: underline;
+}
+
+.advanced-options-content,
+.capabilities-document-content {
+    padding: 1rem;
+    background-color: #f8f9fa;
+    border: 1px solid #dee2e6;
+    border-radius: 4px;
+    margin-top: 0.5rem;
+}
+
+.capabilities-document-content pre {
+    margin: 0;
+    max-height: 400px;
+    overflow: auto;
+    font-size: 0.8rem;
+    background-color: #fff;
+    padding: 1rem;
+    border-radius: 4px;
+}
+
+.preview-map-modal {
+    position: fixed;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    z-index: 1000;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.preview-map-overlay {
+    position: absolute;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background-color: rgba(0, 0, 0, 0.5);
+}
+
+.preview-map-content {
+    position: relative;
+    width: 90%;
+    max-width: 1200px;
+    height: 80%;
+    background-color: #fff;
+    border-radius: 8px;
+    box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
+    display: flex;
+    flex-direction: column;
+    overflow: hidden;
+    z-index: 1001;
+}
+
+.preview-map-header {
+    padding: 1rem 1.5rem;
+    border-bottom: 1px solid #dee2e6;
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    background-color: #f8f9fa;
+}
+
+.preview-map-header h3 {
+    margin: 0;
+    font-size: 1.2rem;
+    color: #333;
+}
+
+.close-button {
+    background: none;
+    border: none;
+    font-size: 1.5rem;
+    cursor: pointer;
+    color: #666;
+    padding: 0;
+    width: 2rem;
+    height: 2rem;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    border-radius: 4px;
+    transition: background-color 0.2s, color 0.2s;
+}
+
+.close-button:hover {
+    background-color: #e9ecef;
+    color: #333;
+}
+
+.preview-map-body {
+    flex: 1;
+    position: relative;
+    overflow: hidden;
+}
+
+.preview-map-container {
+    width: 100%;
+    height: 100%;
+}
