+++ web-app/admin/src/app/admin/admin-layers/layer-details/layer-details.component.ts
@@ -14,7 +14,7 @@ import { SearchModalComponent, SearchModalData, SearchModalResult, SearchModalCo
 import { DeleteLayerComponent } from '../delete-layer/delete-layer.component';
 import { Event } from 'src/app/filter/filter.types';
 import { Observable } from 'rxjs';
-import * as L from 'leaflet';
+import { ImageryLayerConfig } from '../imagery-layer-settings/imagery-layer-settings.component';
 
 interface UrlLayer {
   table: string;
@@ -88,27 +88,19 @@ export class LayerDetailsComponent implements OnInit {
   layerEditForm = {
     name: '',
     description: '',
-    url: '',
     format: '',
-    base: false,
+    base: false
+  };
+
+  // Imagery layer configuration for the helper component
+  imageryConfig: ImageryLayerConfig = {
+    url: '',
+    format: 'XYZ',
     wmsVersion: '1.3.0',
     wmsTransparent: true,
     wmsStyles: ''
   };
-
-  // WMS-related properties
-  wmsCapabilities: any = null;
-  wmsError: string = '';
-  wmsLayers: any[] = [];
-  wmsOtherLayers: any[] = [];
-  selectedWmsLayers: { [key: string]: boolean } = {};
-  advancedOptionsExpanded: boolean = false;
-  isLoadingWms: boolean = false;
-  showPreview: boolean = false;
-  previewMap: L.Map | null = null;
-  previewMapLayer: L.Layer | null = null;
-  showWmsCapabilitiesDocument: boolean = false;
-  wmsLayerSearchQuery: string = '';
+  selectedWmsLayersString: string = '';
 
   @ViewChild('previewMapContainer') previewMapContainer: ElementRef;
 
@@ -390,40 +382,38 @@ export class LayerDetailsComponent implements OnInit {
 
       // Initialize imagery fields if editing an imagery layer
       if (this.layer?.type === 'Imagery') {
-        this.layerEditForm.url = this.layer.url || '';
         this.layerEditForm.format = this.layer.format || 'XYZ';
         this.layerEditForm.base = !!this.layer.base;
 
-        // Initialize WMS fields if WMS layer
-        if (this.layer.format === 'WMS' && this.layer.wms) {
-          this.layerEditForm.wmsVersion = this.layer.wms.version || '1.3.0';
-          this.layerEditForm.wmsTransparent = this.layer.wms.transparent !== false;
-          this.layerEditForm.wmsStyles = this.layer.wms.styles || '';
+        this.imageryConfig = {
+          url: this.layer.url || '',
+          format: this.layer.format || 'XYZ',
+          wmsVersion: this.layer.wms?.version || '1.3.0',
+          wmsTransparent: this.layer.wms?.transparent !== false,
+          wmsStyles: this.layer.wms?.styles || ''
+        };
 
-          // Pre-select layers
-          if (this.layer.wms.layers) {
-            const layers = this.layer.wms.layers.split(',');
-            this.selectedWmsLayers = {};
-            layers.forEach(layerName => {
-              this.selectedWmsLayers[layerName.trim()] = true;
-            });
-          }
-
-          // Auto-fetch capabilities if URL exists
-          if (this.layer.url) {
-            this.fetchWmsCapabilities();
-          }
+        // Store existing WMS layers for pre-selection
+        if (this.layer.format === 'WMS' && this.layer.wms?.layers) {
+          this.selectedWmsLayersString = this.layer.wms.layers;
         }
       }
-    } else {
-      // Reset WMS data when canceling
-      this.wmsCapabilities = null;
-      this.wmsError = '';
-      this.wmsLayers = [];
-      this.wmsOtherLayers = [];
-      this.selectedWmsLayers = {};
     }
     this.editingDetails = !this.editingDetails;
+  }
+
+  /**
+   * Handles imagery config changes from the helper component
+   */
+  onImageryConfigChange(config: ImageryLayerConfig): void {
+    this.imageryConfig = config;
+  }
+
+  /**
+   * Handles WMS layer selection changes from the helper component
+   */
+  onWmsLayersSelected(layers: string): void {
+    this.selectedWmsLayersString = layers;
   }
 
   saveLayerDetails(): void {
@@ -439,20 +429,18 @@ export class LayerDetailsComponent implements OnInit {
 
     // Add imagery-specific fields
     if (this.layer.type === 'Imagery') {
-      updatedLayer.url = this.layerEditForm.url;
-      updatedLayer.format = this.layerEditForm.format;
+      updatedLayer.url = this.imageryConfig.url;
+      updatedLayer.format = this.imageryConfig.format;
       updatedLayer.base = this.layerEditForm.base;
 
       // Add WMS-specific fields
-      if (this.layerEditForm.format === 'WMS') {
-        const selectedLayers = this.getSelectedWmsLayers();
-
+      if (this.imageryConfig.format === 'WMS') {
         updatedLayer.wms = {
-          layers: selectedLayers || '',
-          version: this.layerEditForm.wmsVersion,
-          transparent: this.layerEditForm.wmsTransparent,
-          format: this.layerEditForm.wmsTransparent ? 'image/png' : 'image/jpeg',
-          styles: this.layerEditForm.wmsStyles || ''
+          layers: this.selectedWmsLayersString || '',
+          version: this.imageryConfig.wmsVersion,
+          transparent: this.imageryConfig.wmsTransparent,
+          format: this.imageryConfig.wmsTransparent ? 'image/png' : 'image/jpeg',
+          styles: this.imageryConfig.wmsStyles || ''
         };
       }
     }
@@ -462,11 +450,6 @@ export class LayerDetailsComponent implements OnInit {
         next: (updated) => {
           this.layer = { ...this.layer, ...updated };
           this.editingDetails = false;
-          this.wmsCapabilities = null;
-          this.wmsError = '';
-          this.wmsLayers = [];
-          this.wmsOtherLayers = [];
-          this.selectedWmsLayers = {};
           this.snackBar.open('Layer updated successfully', null, { duration: 2000 });
         },
         error: (error) => {
@@ -481,12 +464,6 @@ export class LayerDetailsComponent implements OnInit {
     this.editingDetails = false;
     this.layerEditForm.name = this.layer?.name || '';
     this.layerEditForm.description = this.layer?.description || '';
-    this.wmsCapabilities = null;
-    this.wmsError = '';
-    this.wmsLayers = [];
-    this.wmsOtherLayers = [];
-    this.selectedWmsLayers = {};
-    this.closePreview();
   }
 
   editLayer(): void {
@@ -685,291 +662,6 @@ export class LayerDetailsComponent implements OnInit {
         setTimeout(() => this.checkLayerProcessingStatus(), 5000);
       }
     });
-  }
-
-  /**
-   * Handles format change when editing imagery layers
-   */
-  onFormatChange(): void {
-    const format = this.layerEditForm.format;
-    const url = this.layerEditForm.url;
-
-    if (format === 'WMS' && url) {
-      this.fetchWmsCapabilities();
-    } else {
-      this.wmsCapabilities = null;
-      this.wmsError = '';
-      this.wmsLayers = [];
-      this.wmsOtherLayers = [];
-      this.selectedWmsLayers = {};
-    }
-  }
-
-  /**
-   * Fetches WMS GetCapabilities document from the server
-   */
-  fetchWmsCapabilities(): void {
-    const url = this.layerEditForm.url;
-    if (!url) {
-      this.wmsError = 'Please enter a WMS URL first';
-      return;
-    }
-
-    this.isLoadingWms = true;
-    this.wmsError = '';
-    this.wmsCapabilities = null;
-    this.wmsLayers = [];
-    this.wmsOtherLayers = [];
-
-    const baseUrl = url.split('?')[0];
-    this.http.post<any>('/api/layers/wms/getcapabilities', { url: baseUrl }).subscribe({
-      next: (response) => {
-        this.isLoadingWms = false;
-        if (response && response.Capability) {
-          this.wmsCapabilities = response;
-          this.parseWmsLayers(response.Capability.Layer, this.wmsLayers, this.wmsOtherLayers);
-          this.layerEditForm.wmsVersion = response.version || '1.3.0';
-
-          if (this.wmsLayers.length === 0 && this.wmsOtherLayers.length === 0) {
-            this.wmsError = 'No layers found in WMS Capabilities document.';
-          }
-        } else {
-          this.wmsError = 'Invalid response received from WMS Server, please check your URL and try again.';
-        }
-      },
-      error: (error) => {
-        this.isLoadingWms = false;
-        let errorMessage = 'Failed to load WMS Capabilities document.';
-
-        if (error.status === 0) {
-          errorMessage = 'Network error: Unable to connect to the WMS server. Please check your connection and URL.';
-        } else if (error.status === 404) {
-          errorMessage = 'WMS server not found (404). Please verify the URL is correct.';
-        } else if (error.status === 500) {
-          errorMessage = 'WMS server error (500). The server may be experiencing issues.';
-        } else if (error.status === 403) {
-          errorMessage = 'Access forbidden (403). You may not have permission to access this WMS server.';
-        } else if (error.error) {
-          if (typeof error.error === 'string') {
-            errorMessage = error.error;
-          } else if (error.error.message) {
-            errorMessage = error.error.message;
-          }
-        }
-
-        this.wmsError = errorMessage;
-      }
-    });
-  }
-
-  /**
-   * Parses WMS layers from capabilities document
-   */
-  private parseWmsLayers(layer: any, layers: any[], otherLayers: any[], layerHierarchy?: string): void {
-    const all = Array.isArray(layer) ? layer : [layer];
-    all.forEach(l => {
-      if (l.Name) {
-        l.Title = layerHierarchy ? `${layerHierarchy} - ${l.Title}` : l.Title;
-        if (this.checkWmsLayer(l)) {
-          layers.push(l);
-        } else {
-          otherLayers.push(l);
-        }
-      }
-
-      if (l.Layer) {
-        this.parseWmsLayers(l.Layer, layers, otherLayers, l.Title);
-      }
-    });
-  }
-
-  /**
-   * Checks if layer supports EPSG:3857
-   */
-  private checkWmsLayer(layer: any): boolean {
-    if (layer.CRS) {
-      return layer.CRS.some((crs: string) =>
-        crs.indexOf('EPSG:3857') !== -1 || crs.indexOf('EPSG:900913') !== -1
-      );
-    }
-    return false;
-  }
-
-  /**
-   * Handles WMS layer selection toggle
-   */
-  onWmsLayerToggle(layerName: string): void {
-    // Toggle is handled by ngModel binding
-  }
-
-  /**
-   * Toggles advanced options visibility
-   */
-  toggleAdvancedOptions(): void {
-    this.advancedOptionsExpanded = !this.advancedOptionsExpanded;
-  }
-
-  /**
-   * Toggles WMS capabilities document visibility
-   */
-  toggleWmsCapabilitiesDocument(): void {
-    this.showWmsCapabilitiesDocument = !this.showWmsCapabilitiesDocument;
-  }
-
-  /**
-   * Opens preview map
-   */
-  openPreview(): void {
-    try {
-      this.showPreview = true;
-      setTimeout(() => {
-        this.initializePreviewMap();
-        this.updatePreviewMap();
-      }, 100);
-    } catch (error) {
-      console.error('Error opening preview:', error);
-      this.snackBar.open('Failed to open preview map. Please try again.', 'Close', { duration: 3000 });
-      this.showPreview = false;
-    }
-  }
-
-  /**
-   * Closes preview map
-   */
-  closePreview(): void {
-    this.showPreview = false;
-    if (this.previewMap) {
-      this.previewMap.remove();
-      this.previewMap = null;
-      this.previewMapLayer = null;
-    }
-  }
-
-  /**
-   * Initializes the Leaflet preview map
-   */
-  private initializePreviewMap(): void {
-    if (!this.previewMapContainer || this.previewMap) return;
-
-    this.previewMap = L.map(this.previewMapContainer.nativeElement, {
-      center: [0, 0],
-      zoom: 3,
-      minZoom: 0,
-      maxZoom: 18,
-      zoomControl: true,
-      trackResize: true,
-      scrollWheelZoom: true,
-      attributionControl: true,
-      worldCopyJump: true
-    });
-
-    L.tileLayer('https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png', {
-      attribution: '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors',
-      maxZoom: 18
-    }).addTo(this.previewMap);
-  }
-
-  /**
-   * Updates the preview map with the current layer configuration
-   */
-  private updatePreviewMap(): void {
-    if (!this.previewMap) return;
-
-    try {
-      if (this.previewMapLayer) {
-        this.previewMap.removeLayer(this.previewMapLayer);
-        this.previewMapLayer = null;
-      }
-
-      const url = this.layerEditForm.url;
-      const format = this.layerEditForm.format;
-
-      if (!url || !format) return;
-
-      if (format === 'XYZ' || format === 'TMS') {
-        this.previewMapLayer = L.tileLayer(url, {
-          tms: format === 'TMS',
-          maxZoom: 18,
-          errorTileUrl: 'data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAABCAYAAAAfFcSJAAAADUlEQVR42mNk+M9QDwADhgGAWjR9awAAAABJRU5ErkJggg=='
-        }).addTo(this.previewMap);
-      } else if (format === 'WMS') {
-        const selectedLayers = this.getSelectedWmsLayers();
-        if (!selectedLayers) return;
-
-        const wmsOptions: any = {
-          layers: selectedLayers,
-          version: this.layerEditForm.wmsVersion || '1.3.0',
-          format: this.layerEditForm.wmsTransparent ? 'image/png' : 'image/jpeg',
-          transparent: this.layerEditForm.wmsTransparent
-        };
-
-        if (this.layerEditForm.wmsStyles) {
-          wmsOptions.styles = this.layerEditForm.wmsStyles;
-        }
-
-        this.previewMapLayer = L.tileLayer.wms(url, wmsOptions).addTo(this.previewMap);
-
-        const firstSelectedLayer = this.wmsLayers.find(l => this.selectedWmsLayers[l.Name]);
-        if (firstSelectedLayer?.EX_GeographicBoundingBox) {
-          const extent = firstSelectedLayer.EX_GeographicBoundingBox;
-          const bounds = L.latLngBounds(
-            [extent[1], extent[0]],
-            [extent[3], extent[2]]
-          );
-          this.previewMap.fitBounds(bounds);
-        }
-      }
-
-      setTimeout(() => {
-        if (this.previewMap) {
-          this.previewMap.invalidateSize();
-        }
-      }, 100);
-    } catch (error) {
-      console.error('Error updating preview map:', error);
-      this.snackBar.open('Failed to load layer in preview. Please check your configuration.', 'Close', { duration: 3000 });
-    }
-  }
-
-  /**
-   * Filters WMS layers based on search query
-   */
-  filteredWmsLayers(): any[] {
-    if (!this.wmsLayerSearchQuery || this.wmsLayerSearchQuery.trim() === '') {
-      return this.wmsLayers;
-    }
-
-    const query = this.wmsLayerSearchQuery.toLowerCase();
-    return this.wmsLayers.filter(layer =>
-      layer.Title?.toLowerCase().includes(query) ||
-      layer.Name?.toLowerCase().includes(query) ||
-      layer.Abstract?.toLowerCase().includes(query)
-    );
-  }
-
-  /**
-   * Filters unavailable WMS layers based on search query
-   */
-  filteredWmsOtherLayers(): any[] {
-    if (!this.wmsLayerSearchQuery || this.wmsLayerSearchQuery.trim() === '') {
-      return this.wmsOtherLayers;
-    }
-
-    const query = this.wmsLayerSearchQuery.toLowerCase();
-    return this.wmsOtherLayers.filter(layer =>
-      layer.Title?.toLowerCase().includes(query) ||
-      layer.Name?.toLowerCase().includes(query) ||
-      layer.Abstract?.toLowerCase().includes(query)
-    );
-  }
-
-  /**
-   * Gets selected WMS layer names as comma-separated string
-   */
-  getSelectedWmsLayers(): string {
-    return Object.keys(this.selectedWmsLayers)
-      .filter(name => this.selectedWmsLayers[name])
-      .join(',');
   }
 }
